requirements.txt
Django==5.2.6
djangorestframework==3.16.1
djongo==1.2.31
pymongo==4.14.1
djangorestframework-simplejwt==5.5.1
python-decouple==3.8
django-cors-headers==4.8.0
dnspython==2.8.0